Currency Converter Project
Chuck-a-luck is played with three standard dice that are kept in a device shaped somewhat like an hourglass that resembles
a wire-frame bird cage and pivots about its center. The dealer rotates the cage end over end, with the dice landing on the bottom.

Wagers are placed based on possible combinations that can appear on the three dice. 
The possible wagers are usually fewer than the wagers that are possible in sic bo and, in that sense, 
chuck-a-luck can be considered to be a simpler game.

The wagers, and their associated odds, that are typically available are set out in the table below.

Type	                                  Wager	                                                                              Odds
Single Die Bet	                        A specific number will appear	                                                  1 die, 1 to 1; 
.                                       .                                                                               2 dice, 2 to 1;
.                                       .                                                                               3 dice,10 to 1
.                                       .                                                                          (sometimes 3 to 1)
Any Triple (sometimes offered)	        Any of the triples (all three dice show the same number) will appear	              30 to 1
Big (sometimes offered)	                The total score will be 11 (sometimes 12) or higher with the exception of a triple	1 to 1
Small (sometimes offered)	              The total score will be 10 (sometimes 9) or lower with the exception of a triple	  1 to 1
Field (sometimes offered)	              The total score will be outside the range of 8 to 12 (inclusive)	                  1 to 1


Chuck-a-luck is a game of chance. On average, the players are expected to lose more than they win. 
The casino's advantage (house advantage or house edge) is greater than most other casino games and can be much greater.

For example, there are 216 (6 × 6 × 6) possible outcomes for a single throw of three dice. For a specific number:

There are 75 possible outcomes, where only one die will match the number;
There are 15 possible outcomes, where only two dice will match; and
There is one possible outcome, where all three dice will match.
At odds of 1 to 1, 2 to 1 and 10 to 1 respectively for each of these types of outcome, 
the expected loss as a percentage of the stake wagered is:

1 - ((75/216) × 2 + (15/216) × 3 + (1/216) × 11) = 4.6%

At worse odds of 1 to 1, 2 to 1 and 3 to 1, the expected loss as a percentage of the stake wagered is:

1 - ((75/216) × 2 + (15/216) × 3 + (1/216) × 4) = 7.9%

If the odds are adjusted to 1 to 1, 3 to 1 and 5 to 1 respectively, the expected loss as a percentage is:

1 - ((75/216) × 2 + (15/216) × 4 + (1/216) × 6) = 0%

However, commercially organised gambling games always have a house advantage which acts as a fee for the privilege
of being allowed to play the game, so the last scenario does not represent real practice.